requirements.txt
langchain==0.2.14
langchain-openai==0.1.22
python-dotenv==1.0.1
